+++ tests/tests/test-02-mirtTwo.R
@@ -137,6 +137,7 @@ test_that('poly', {
     expect_equal(gM2$M2, 6.518739, tolerance = 1e-4)
     expect_equal(gM2$df, 5)
     expect_equal(gM2$CFI, .9961693, tolerance = 1e-4)
+    expect_equal(gM2$SRMSR, 0.01691436, tolerance = 1e-4)
 
     #item and test info
     Theta <- matrix(seq(-4,4,.01))